requirements.txt
apify-client
requests
pymongo
python-dotenv
yfinance
certifi
pandas
